backend_python/requirements.txt
fastapi
uvicorn
pymssql
boto3
bcrypt
PyJWT
python-dotenv
pydantic